+++ libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/25 19:15:07 by auquinta          #+#    #+#             */
-/*   Updated: 2021/10/28 17:13:42 by jvacaris         ###   ########.fr       */
+/*   Updated: 2021/11/13 12:30:34 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,5 +71,5 @@ t_list	*ft_lstlast(t_list *lst);
 int		get_next_line(int fd, char **line);
 int		is_valid_var(char prv_char, char curr_char, char nxt_char, char qm);
 char	*qm_remover(char *qm_str);
-
+char	*ft_strreplace(const char *str, char const *oldset, char const *newset);
 #endif